requirements.txt
numpy~=1.22.1
opencv-contrib-python
pyautogui~=0.9.53
keyboard~=0.13.5
pillow~=9.0.0
pyscreenshot~=3.0
scipy~=1.7.3
pytesseract
imutils~=0.5.4
pandas~=1.4.0
plotly~=5.5.0
requests~=2.28.1
pynput~=1.7.6
screeninfo~=0.8.1